requirements.txt
matplotlib==3.10.8
numpy==2.4.1
PyYAML==6.0.3
Requests==2.32.5
tqdm==4.67.1
